Moving config file parsing from app to CLI
