+++ src/helpers/configUtils.js
@@ -1,7 +1,19 @@
+const fs = require('fs');
+const path = require('path');
 const Ajv = require('ajv');
 const metaSchema = require('ajv/lib/refs/json-schema-draft-06.json');
 const logger = require('./logger');
 const configSchema = require('./schemas/config.schema.json');
+
+function getConfig(pathToConfig) {
+  // Checks pathToConfig points to valid JSON file
+  const fullPath = path.resolve(pathToConfig);
+  try {
+    return JSON.parse(fs.readFileSync(fullPath));
+  } catch (err) {
+    throw new Error(`The provided filepath to a configuration file ${pathToConfig}, full path ${fullPath} did not point to a valid JSON file.`);
+  }
+}
 
 const ajv = new Ajv({ logger: false, allErrors: true });
 ajv.addMetaSchema(metaSchema);
@@ -34,5 +46,6 @@ function validateConfig(config) {
 }
 
 module.exports = {
+  getConfig,
   validateConfig,
 };
